bug: set project via a flag if specified (#209)
Because /auth exports the project ID as an environment variable, that environment variable takes precedence over the core property. This previously worked because setup-gcloud didn't export said environment variable. To fix this, I updated the gcloud command to use the --project flag if a project ID was given. This has the added side-effect of being able to remove all the code to configure gcloud, since it was never really necessary.

Also, I'm not sure it was ever a good idea to modify global configuration for gcloud just because someone deployed a service. Using the flag is probably the best approach here. You might notice that I added PROJECT_ID to the e2e tests. That's because they were secretly tightly coupled with the previous step that had previously run.